+++ android/src/main/java/com/reactlibrary/sound/SoundPlayer.java
@@ -1,7 +1,7 @@
 package com.reactlibrary.sound;
 
 public interface SoundPlayer {
-    void playNote(final int note, long duration);
+    void playNote(final int note, int duration);
 
     void noteOn(int note);
